# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FleetSpecial API
  version: 1.0.0
  description: |
    Contrato **REST (API First)** de FleetSpecial — SaaS multi-tenant para transporte
    especial y flotas pequeñas en Colombia.

    Este contrato **antecede al código** (Spec Driven Development): de él se generan el
    SDK del portal web, los mocks para que web y móvil avancen en paralelo, y la
    validación de requests en runtime. Deriva de las specs aprobadas (Fase 3) y del
    modelo de dominio (Fase 2).

    ## Trazabilidad spec → endpoints
    - spec-001 Onboarding de Empresa (Tenant) ....... POST /tenants
    - spec-002 Invitar usuarios y roles ............. POST /usuarios, /usuarios/{id}
    - spec-003 Registrar Vehículo ................... /vehiculos
    - spec-004 Registrar Conductor + Licencia ....... /conductores
    - spec-005 Registrar Documento + adjunto ........ /documentos
    - spec-006 Alertas / Semáforo ................... GET /cumplimiento/*
    - spec-007 Renovación de Documento .............. POST /documentos/{id}/renovaciones
    - spec-008 Crear Servicio + Asignación .......... /servicios, /asignaciones
    - spec-009 Regla de oro (bloqueo asignación) .... efecto en POST /asignaciones (409)
    - spec-010 Conductor offline (mi día, sync) ..... /sync/*, /servicios/{id}/estado
    - spec-011 Tanqueo offline append-only .......... POST /combustible (Idempotency-Key)

    ## Convenciones (Fase 5)
    - Versionado por **prefijo de ruta** (`/v1`).
...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /vehiculos/{vehiculoId}:
    parameters:
      - $ref: '#/components/parameters/VehiculoId'
    get:
      tags: [Vehículos]
      summary: Obtener un vehículo
      operationId: obtenerVehiculo
      responses:
        '200':
          description: Vehículo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehiculo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Vehículos]
      summary: Actualizar datos básicos del vehículo
      operationId: actualizarVehiculo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActualizarVehiculoRequest' }
      responses:
        '200':
          description: Vehículo actualizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehiculo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /conductores/{conductorId}:
    parameters:
      - $ref: '#/components/parameters/ConductorId'
    get:
      tags: [Conductores]
      summary: Obtener un conductor
      operationId: obtenerConductor
      responses:
        '200':
          description: Conductor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conductor' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────── spec-005 R2/R10 Catálogo de Tipos de documento ─────────────────
  /catalogo/tipos:
    get:
      tags: [Catálogo]
      summary: Listar los Tipos de documento del tenant
      operationId: listarTiposDocumento
      responses:
        '200':
          description: Tipos del catálogo (activos e inactivos).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TipoDocumentoCatalogo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Catálogo]
      summary: Agregar un Tipo de documento al catálogo
      description: |
        spec-005 R2/R10. El catálogo es configurable por tenant sin redeploy:
        cambios normativos se resuelven agregando o desactivando Tipos.
      operationId: agregarTipoDocumento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgregarTipoDocumentoRequest' }
      responses:
        '201':
          description: Tipo agregado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TipoDocumentoCatalogo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Ya existe un Tipo con ese código en el tenant.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /catalogo/tipos/{codigo}:
    parameters:
      - name: codigo
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Catálogo]
      summary: Actualizar un Tipo (activar/desactivar, requerido)
      description: |
        spec-005 R10. Desactivar un Tipo impide registrar nuevos Documentos de ese
        Tipo sin afectar los existentes. `requerido` alimenta la invariante I3
        (requerido ausente ⇒ Semáforo rojo).
      operationId: actualizarTipoDocumento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActualizarTipoDocumentoRequest' }
      responses:
        '200':
          description: Tipo actualizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TipoDocumentoCatalogo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────── spec-005 / 007 Documentos ─────────────────────────
  /documentos:
    get:
      tags: [Documentos]
      summary: Listar documentos (filtrable por sujeto)
      operationId: listarDocumentos
      parameters:
        - name: sujetoTipo
          in: query
          schema: { type: string, enum: [vehiculo, conductor] }
        - name: sujetoId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página de documentos.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagina'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Documento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Documentos]
      summary: Registrar un documento con vencimiento
      description: |
        spec-005. Asocia un Documento a un Vehículo o Conductor con su fecha de
        vencimiento. El binario del adjunto se sube por separado vía
        `POST /documentos/{id}/adjunto`. Emite `DocumentoRegistrado`.
      operationId: registrarDocumento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistrarDocumentoRequest' }
      responses:
        '201':
          description: Documento creado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Documento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /documentos/{documentoId}:
    parameters:
      - $ref: '#/components/parameters/DocumentoId'
    get:
      tags: [Documentos]
      summary: Obtener un documento (con histórico de versiones)
      operationId: obtenerDocumento
      responses:
        '200':
          description: Documento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Documento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /documentos/{documentoId}/renovaciones:
    parameters:
      - $ref: '#/components/parameters/DocumentoId'
    post:
      tags: [Documentos]
      summary: Renovar un documento conservando histórico
      description: |
        spec-007. Crea una nueva versión vigente y conserva la anterior en el histórico.
        Emite `DocumentoRenovado`; si el sujeto vuelve a estar al día, rehabilita
        asignaciones (política P5).
      operationId: renovarDocumento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenovarDocumentoRequest' }
      responses:
        '201':
          description: Documento renovado (nueva versión vigente).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Documento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ───────────────────────── spec-006 Cumplimiento / Semáforo ─────────────────────────
  /cumplimiento/vehiculos/{vehiculoId}:
    parameters:
      - $ref: '#/components/parameters/VehiculoId'
    get:
      tags: [Cumplimiento]
      summary: Estado de cumplimiento (Semáforo) de un vehículo
      description: spec-006. Devuelve el Semáforo y el detalle por documento.
      operationId: cumplimientoVehiculo
      responses:
        '200':
          description: Estado de cumplimiento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstadoCumplimiento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /cumplimiento/conductores/{conductorId}:
    parameters:
      - $ref: '#/components/parameters/ConductorId'
    get:
      tags: [Cumplimiento]
      summary: Estado de cumplimiento (Semáforo) de un conductor
      operationId: cumplimientoConductor
      responses:
        '200':
          description: Estado de cumplimiento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstadoCumplimiento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /cumplimiento/alertas:
    get:
      tags: [Cumplimiento]
      summary: Listar alertas de vencimiento vigentes
      description: |
        spec-006. Documentos por vencer (30/15/3 días) y vencidos del tenant.
        Alimenta el tablero de alertas del portal.
      operationId: listarAlertas
      parameters:
        - name: estado
          in: query
          schema: { type: string, enum: [por_vencer, vencido] }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página de alertas.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagina'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Alerta' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────── spec-008 / 009 / 010 Servicios ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT OIDC. El token incluye el claim `tenant_id` y los `roles`. El servidor
        deriva el tenant del token (nunca del body/query) y lo fija en la sesión de
        base de datos para que Row Level Security aísle los datos (ADR-0008).
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    VehiculoId:
      name: vehiculoId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ConductorId:
      name: conductorId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DocumentoId:
      name: documentoId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Falta o es inválido el token de autenticación.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NotFound:
      description: Recurso no encontrado en el tenant actual.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Error de validación de la solicitud.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: Error con forma RFC 7807.
      properties:
        type: { type: string, description: 'Identificador del tipo de error (p. ej. conflicto_horario, incumplimiento).' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Detalle de errores de validación por campo.
          items:
            type: object
            properties:
              campo: { type: string }
              mensaje: { type: string }
      required: [title, status]
    Pagina:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
      required: [page, pageSize, total]
...
    Semaforo:
      type: string
      description: Estado de cumplimiento. Verde/Amarillo/Rojo.
      enum: [Vigente, PorVencer, Vencido]
    SujetoRef:
      type: object
      description: Referencia polimórfica a un Vehículo o Conductor.
      properties:
        tipo: { type: string, enum: [vehiculo, conductor] }
        id: { type: string, format: uuid }
      required: [tipo, id]

    # ---------- Tenant / Usuario ----------
...
    ClaseVehiculo:
      type: string
      enum: [automovil, camioneta, van, microbus, bus, campero, otro]
...
    ActualizarVehiculoRequest:
      type: object
      properties:
        marca: { type: string }
        modelo: { type: string }
        anio: { type: integer }
        propietarioId: { type: string, format: uuid }
        estado: { type: string, enum: [activo, inactivo] }
    Vehiculo:
      type: object
      properties:
        id: { type: string, format: uuid }
        placa: { type: string }
        clase: { $ref: '#/components/schemas/ClaseVehiculo' }
        marca: { type: string }
        modelo: { type: string }
        anio: { type: integer }
        odometro: { type: integer, description: Última lectura conocida. }
        semaforo: { $ref: '#/components/schemas/Semaforo' }
        estado: { type: string, enum: [activo, inactivo] }
      required: [id, placa, clase, estado]
...
    Conductor:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        usuarioId: { type: string, format: uuid }
        licencia:
          type: object
          properties:
            numero: { type: string }
            categoria: { type: string }
            vencimiento: { type: string, format: date }
        semaforo: { $ref: '#/components/schemas/Semaforo' }
      required: [id, nombre]

    # ---------- Catálogo de Tipos (spec-005 R2/R10) ----------
    TipoDocumentoCatalogo:
      type: object
      properties:
        codigo: { type: string, description: 'P. ej. SOAT, RTM, LICENCIA.' }
        aplicaA: { type: string, enum: [vehiculo, conductor] }
        requerido: { type: boolean, description: 'Alimenta la invariante I3 (ausente ⇒ rojo).' }
        activo: { type: boolean }
      required: [codigo, aplicaA, requerido, activo]
    AgregarTipoDocumentoRequest:
      type: object
      properties:
        codigo: { type: string }
        aplicaA: { type: string, enum: [vehiculo, conductor] }
        requerido: { type: boolean, default: false }
      required: [codigo, aplicaA]
    ActualizarTipoDocumentoRequest:
      type: object
      description: Campos opcionales; solo se actualiza lo enviado.
      properties:
        activo: { type: boolean }
        requerido: { type: boolean }

    # ---------- Documento ----------
    RegistrarDocumentoRequest:
      type: object
      properties:
        sujeto: { $ref: '#/components/schemas/SujetoRef' }
        tipo: { type: string, description: 'Código del Tipo de documento (catálogo configurable): SOAT, RTM, TARJETA_OPERACION, LICENCIA, EXAMEN_MEDICO, POLIZA_CONTRACTUAL, POLIZA_EXTRACONTRACTUAL, etc.' }
        numero: { type: string }
        expedicion: { type: string, format: date }
        vencimiento: { type: string, format: date }
      required: [sujeto, tipo, vencimiento]
    RenovarDocumentoRequest:
      type: object
      properties:
        numero: { type: string }
        expedicion: { type: string, format: date }
        vencimiento: { type: string, format: date }
      required: [vencimiento]
    Documento:
      type: object
      properties:
        id: { type: string, format: uuid }
        sujeto: { $ref: '#/components/schemas/SujetoRef' }
        tipo: { type: string }
        numero: { type: string }
        expedicion: { type: string, format: date }
        vencimiento: { type: string, format: date }
        estado: { $ref: '#/components/schemas/Semaforo' }
        tieneAdjunto: { type: boolean }
        version: { type: integer, description: Número de versión (renovaciones). }
        historico:
          type: array
          description: Versiones anteriores (renovaciones).
          items:
            type: object
            properties:
              version: { type: integer }
              vencimiento: { type: string, format: date }
              reemplazadoEn: { type: string, format: date-time }
      required: [id, sujeto, tipo, vencimiento, estado]

    # ---------- Cumplimiento ----------
    EstadoCumplimiento:
      type: object
      properties:
        sujeto: { $ref: '#/components/schemas/SujetoRef' }
        semaforo: { $ref: '#/components/schemas/Semaforo' }
        documentos:
          type: array
          items:
            type: object
            properties:
              documentoId: { type: string, format: uuid }
              tipo: { type: string }
              vencimiento: { type: string, format: date }
              estado: { $ref: '#/components/schemas/Semaforo' }
              diasRestantes: { type: integer }
      required: [sujeto, semaforo]
    Alerta:
      type: object
      properties:
        documentoId: { type: string, format: uuid }
        sujeto: { $ref: '#/components/schemas/SujetoRef' }
        tipo: { type: string }
        estado: { type: string, enum: [por_vencer, vencido] }
        vencimiento: { type: string, format: date }
        diasRestantes: { type: integer, description: 'Negativo si ya venció.' }
      required: [documentoId, sujeto, estado, vencimiento]

    # ---------- Servicio ----------
...